+++ tests/unit/components/array-inline-item-test.js
@@ -16,6 +16,7 @@ describeComponent(
       onChange: PropTypes.func.isRequired,
       onRemove: PropTypes.func.isRequired,
       readOny: PropTypes.bool,
+      sortable: PropTypes.bool.isRequired,
       store: PropTypes.EmberObject.isRequired,
       value: PropTypes.object.isRequired
     })