+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-#5.1
+# 5.2
+
+## Non-Breaking Changes
+* **Added** support for sortable arrays. You can use this feature by adding `sortable` in your view for the array item container.
+
+```json
+…
+{
+  "model": "addresses",
+  "item": {
+    "container": "address",
+    "sortable": true
+  }
+}
+…
+```
+
+# 5.1
 
 ## Non-Breaking Changes
 * **Added** support for arrays that automatically add an empty item to the end of the list instead of using the `Add Item` button. You can use this feature by adding `autoAdd` in your view for the array item container.